Move states current / desired discovery tests to `states_discover_cmd` tests.
