+++ crate/rt/src/cmds/sub.rs
@@ -7,14 +7,10 @@
 
 pub use self::{
     apply_cmd::{ApplyCmd, ApplyFor},
-    states_current_discover_cmd::StatesCurrentDiscoverCmd,
-    states_desired_discover_cmd::StatesDesiredDiscoverCmd,
     states_desired_read_cmd::StatesDesiredReadCmd,
     states_saved_read_cmd::StatesSavedReadCmd,
 };
 
 mod apply_cmd;
-mod states_current_discover_cmd;
-mod states_desired_discover_cmd;
 mod states_desired_read_cmd;
 mod states_saved_read_cmd;
