+++ workspace_tests/src/rt/cmds/states_discover_cmd.rs
@@ -3,7 +3,7 @@ use peace::{
     cmd::ctx::CmdCtx,
     resources::{
         paths::{StatesDesiredFile, StatesSavedFile},
-        states::{StatesCurrent, StatesDesired},
+        states::{StatesCurrent, StatesDesired, StatesSaved},
         type_reg::untagged::{BoxDtDisplay, TypeReg},
     },
     rt::cmds::StatesDiscoverCmd,
@@ -13,7 +13,8 @@ use peace::{
 use crate::{NoOpOutput, PeaceTestError, VecCopyError, VecCopyItemSpec, VecCopyState};
 
 #[tokio::test]
-async fn runs_state_current_and_state_desired() -> Result<(), Box<dyn std::error::Error>> {
+async fn current_and_desired_discovers_both_states_current_and_desired()
+-> Result<(), Box<dyn std::error::Error>> {
     let tempdir = tempfile::tempdir()?;
     let workspace = Workspace::new(
         app_name!(),
@@ -62,6 +63,145 @@ async fn runs_state_current_and_state_desired() -> Result<(), Box<dyn std::error
         states_current.get::<VecCopyState, _>(VecCopyItemSpec.id()),
         states_on_disk.get::<VecCopyState, _>(VecCopyItemSpec.id())
     );
+    assert_eq!(
+        Some(VecCopyState::from(vec![0u8, 1, 2, 3, 4, 5, 6, 7])).as_ref(),
+        vec_copy_desired_state
+    );
+    assert_eq!(
+        states_desired.get::<VecCopyState, _>(VecCopyItemSpec.id()),
+        states_desired_on_disk.get::<VecCopyState, _>(VecCopyItemSpec.id())
+    );
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn current_runs_state_current_for_each_item_spec() -> Result<(), Box<dyn std::error::Error>> {
+    let tempdir = tempfile::tempdir()?;
+    let workspace = Workspace::new(
+        app_name!(),
+        WorkspaceSpec::Path(tempdir.path().to_path_buf()),
+    )?;
+    let graph = {
+        let mut graph_builder = ItemSpecGraphBuilder::<PeaceTestError>::new();
+        graph_builder.add_fn(VecCopyItemSpec.into());
+        graph_builder.build()
+    };
+    let flow = Flow::new(FlowId::new(crate::fn_name_short!())?, graph);
+    let mut output = NoOpOutput;
+    let mut cmd_ctx = CmdCtx::builder_single_profile_single_flow(&mut output, &workspace)
+        .with_profile(profile!("test_profile"))
+        .with_flow(&flow)
+        .await?;
+
+    let states_current = StatesDiscoverCmd::current(&mut cmd_ctx).await?;
+    let resources = cmd_ctx.resources();
+
+    let vec_copy_state = states_current.get::<VecCopyState, _>(VecCopyItemSpec.id());
+    let states_on_disk = {
+        let states_saved_file = resources.borrow::<StatesSavedFile>();
+        let states_slice = std::fs::read(&*states_saved_file)?;
+
+        let mut type_reg = TypeReg::<ItemSpecId, BoxDtDisplay>::new_typed();
+        type_reg.register::<VecCopyState>(VecCopyItemSpec.id().clone());
+
+        let deserializer = serde_yaml::Deserializer::from_slice(&states_slice);
+        StatesCurrent::from(type_reg.deserialize_map(deserializer)?)
+    };
+    assert_eq!(Some(VecCopyState::new()).as_ref(), vec_copy_state);
+    assert_eq!(
+        states_current.get::<VecCopyState, _>(VecCopyItemSpec.id()),
+        states_on_disk.get::<VecCopyState, _>(VecCopyItemSpec.id())
+    );
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn current_inserts_states_saved_from_states_saved_file()
+-> Result<(), Box<dyn std::error::Error>> {
+    let tempdir = tempfile::tempdir()?;
+    let workspace = Workspace::new(
+        app_name!(),
+        WorkspaceSpec::Path(tempdir.path().to_path_buf()),
+    )?;
+    let graph = {
+        let mut graph_builder = ItemSpecGraphBuilder::<PeaceTestError>::new();
+        graph_builder.add_fn(VecCopyItemSpec.into());
+        graph_builder.build()
+    };
+    let flow = Flow::new(FlowId::new(crate::fn_name_short!())?, graph);
+    let mut output = NoOpOutput;
+    let mut cmd_ctx = CmdCtx::builder_single_profile_single_flow(&mut output, &workspace)
+        .with_profile(profile!("test_profile"))
+        .with_flow(&flow)
+        .await?;
+
+    // Writes to states_saved_file.yaml
+    StatesDiscoverCmd::current(&mut cmd_ctx).await?;
+
+    // Execute again to ensure StatesSaved is included
+    let mut cmd_ctx = CmdCtx::builder_single_profile_single_flow(&mut output, &workspace)
+        .with_profile(profile!("test_profile"))
+        .with_flow(&flow)
+        .await?;
+    StatesDiscoverCmd::current(&mut cmd_ctx).await?;
+    let resources = cmd_ctx.resources();
+
+    let states_saved = resources.borrow::<StatesSaved>();
+    let vec_copy_state = states_saved.get::<VecCopyState, _>(VecCopyItemSpec.id());
+    let states_on_disk = {
+        let states_saved_file = resources.borrow::<StatesSavedFile>();
+        let states_slice = std::fs::read(&*states_saved_file)?;
+
+        let mut type_reg = TypeReg::<ItemSpecId, BoxDtDisplay>::new_typed();
+        type_reg.register::<VecCopyState>(VecCopyItemSpec.id().clone());
+
+        let deserializer = serde_yaml::Deserializer::from_slice(&states_slice);
+        StatesCurrent::from(type_reg.deserialize_map(deserializer)?)
+    };
+    assert_eq!(Some(VecCopyState::new()).as_ref(), vec_copy_state);
+    assert_eq!(
+        states_saved.get::<VecCopyState, _>(VecCopyItemSpec.id()),
+        states_on_disk.get::<VecCopyState, _>(VecCopyItemSpec.id())
+    );
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn desired_runs_state_desired_for_each_item_spec() -> Result<(), Box<dyn std::error::Error>> {
+    let tempdir = tempfile::tempdir()?;
+    let workspace = Workspace::new(
+        app_name!(),
+        WorkspaceSpec::Path(tempdir.path().to_path_buf()),
+    )?;
+    let graph = {
+        let mut graph_builder = ItemSpecGraphBuilder::<PeaceTestError>::new();
+        graph_builder.add_fn(VecCopyItemSpec.into());
+        graph_builder.build()
+    };
+    let flow = Flow::new(FlowId::new(crate::fn_name_short!())?, graph);
+    let mut output = NoOpOutput;
+    let mut cmd_ctx = CmdCtx::builder_single_profile_single_flow(&mut output, &workspace)
+        .with_profile(profile!("test_profile"))
+        .with_flow(&flow)
+        .await?;
+
+    let states_desired = StatesDiscoverCmd::desired(&mut cmd_ctx).await?;
+    let resources = cmd_ctx.resources();
+
+    let vec_copy_desired_state = states_desired.get::<VecCopyState, _>(VecCopyItemSpec.id());
+    let states_desired_on_disk = {
+        let states_desired_file = resources.borrow::<StatesDesiredFile>();
+        let states_slice = std::fs::read(&*states_desired_file)?;
+
+        let mut type_reg = TypeReg::<ItemSpecId, BoxDtDisplay>::new_typed();
+        type_reg.register::<VecCopyState>(VecCopyItemSpec.id().clone());
+
+        let deserializer = serde_yaml::Deserializer::from_slice(&states_slice);
+        StatesDesired::from(type_reg.deserialize_map(deserializer)?)
+    };
     assert_eq!(
         Some(VecCopyState::from(vec![0u8, 1, 2, 3, 4, 5, 6, 7])).as_ref(),
         vec_copy_desired_state
